webapp/pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "A Python-based AI chatbot using FastAPI and OpenAI's API"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "openai",
    "python-dotenv",
    "pydantic",
    "pyyaml",
    "jinja2",
    "grpcio>=1.70.0",
    "grpcio-tools>=1.70.0",
    "websockets>=13.1",
    "websocket-client>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
required-version = ">=0.1.0"
